datasimulateforiot.py
uploaded data to act as an emulation of an IoT device in replacement for using the raspberry pi emulator. Can also copy and paste content on IoT device emulators should they be running python. (Note: this is python code, may require downloading python and compiling this code via the cmb to make it operate.